+++ spec/unit/installer_spec.rb
@@ -113,8 +113,8 @@ module Pod
         end
 
         it "configures the analyzer to use update mode if appropriate" do
-          @installer.update_mode = true
-          Installer::Analyzer.any_instance.expects(:update_mode=).with(true)
+          @installer.update = true
+          Installer::Analyzer.any_instance.expects(:update=).with(true)
           @installer.send(:analyze)
           @installer.aggregate_targets.map(&:name).sort.should == ['Pods']
           @installer.pod_targets.map(&:name).sort.should == ['Pods-JSONKit']